module: add hi3516dv300 clk driver
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I5XHIQ

-----------------------------------------------------------

Add clk driver support to series socs including
3516dv300,3519av100,etc.

+++ drivers/clk/hisilicon/clk-hisi-phase.c
@@ -77,7 +77,7 @@ static int hisi_clk_set_phase(struct clk_hw *hw, int degrees)
 
 	val = readl(phase->reg);
 	val &= ~phase->mask;
-	val |= regval << phase->shift;
+	val |= (unsigned int)regval << phase->shift;
 	writel(val, phase->reg);
 
 	spin_unlock_irqrestore(phase->lock, flags);
